requirements.txt
django>=3.1.6
requests==2.25.1
django-phonenumber-field==5.0.0